work for the https://github.com/surveyjs/private-tasks/issues/475
