+++ packages/survey-creator-core/src/drag-drop-enums.ts
@@ -1,0 +1,15 @@
+export enum DropIndicatorPosition {
+  Inside = "inside",
+  Top = "top",
+  Bottom = "bottom",
+  Right = "right",
+  Left = "left"
+}
+
+export enum ElType {
+  Page = "page",
+  Panel = "panel",
+  Question = "question",
+  DynamicPanel = "dynamic-panel",
+  EmptySurvey = "empty-survey"
+}
