--- packages/survey-creator-core/src/drop-to-enum.ts
@@ -1,7 +1,0 @@
-export enum DropIndicatorPosition {
-  Inside = "inside",
-  Top = "top",
-  Bottom = "bottom",
-  Right = "right",
-  Left = "left"
-}
