+++ packages/survey-creator-core/tests/components.tests.ts
@@ -10,7 +10,7 @@ import { CreatorTester } from "./creator-tester";
 import { LogoImageViewModel } from "../src/components/header/logo-image";
 import { imageMimeTypes } from "../src/utils/utils";
 import { calculateDragOverLocation } from "../src/dragdrop-survey-elements";
-import { DropIndicatorPosition } from "../src/drop-to-enum";
+import { DropIndicatorPosition } from "../src/drag-drop-enums";
 
 beforeEach(() => { });
