+++ packages/survey-creator-core/tests/dragdrop.tests.ts
@@ -1,6 +1,6 @@
 import { SurveyModel, QuestionTextModel } from "survey-core";
 import { DragDropSurveyElements } from "../src/dragdrop-survey-elements";
-import { DropIndicatorPosition } from "../src/drop-to-enum";
+import { DropIndicatorPosition } from "../src/drag-drop-enums";
 import { CreatorTester } from "./creator-tester";
 import { QuestionAdornerViewModel } from "../src/components/question";
